+++ src/task2/my_malloc.h
@@ -2,11 +2,12 @@
 #ifndef _MY_MALLOC_H_
 #define _MY_MALLOC_H_
 
+#include <stdint.h>
 #include <stdlib.h>
 
-void* my_malloc(size_t size);
+void* my_malloc(uint64_t size);
 void my_free(void* ptr);
-void my_allocator_init(size_t size);
+void my_allocator_init(uint64_t size);
 void my_allocator_destroy(void);
 
 #endif
