+++ src/tests/membench.h
@@ -2,11 +2,12 @@
 #define MEMBENCH_H
 
 #include <stddef.h>
+#include <stdint.h>
 
-typedef void (*init_allocator_fn)(size_t);
+typedef void (*init_allocator_fn)(uint64_t);
 typedef void (*destroy_allocator_fn)(void);
 
-typedef void* (*malloc_fn)(size_t);
+typedef void* (*malloc_fn)(uint64_t);
 typedef void (*free_fn)(void*);
 
 /**
